requirements.txt
cmake==3.22.4
dlib==19.24.0
face-recognition==1.3.0
numpy==1.23.2
typing_extensions==4.1.1
protobuf~=3.19.0
aiortc==1.3.2
matplotlib==3.5.1
pydub==0.25.
opencv_python==4.6.0.66
pandas==1.4.4
Pillow==9.2.0
Pyrebase4==4.5.0
streamlit==1.12.2
streamlit_authenticator==0.2.1
streamlit_option_menu==0.3.2
streamlit_webrtc==0.43.4



